fix: atm page
