+++ templates/fossa/atm_receipt.html
@@ -134,7 +134,7 @@
     el: '#vue',
     mixins: [windowMixin],
     delimiters: ['${', '}'],
-    data: function () {
+    data() {
       return {
         theurl: location.protocol + '//' + location.host,
         printDialog: {
