+++ templates/fossa/index.html
@@ -170,7 +170,7 @@
                   flat
                   dense
                   size="xs"
-                  @click="openATMLink(props.row.deviceid, props.row.payload)"
+                  @click="openAtmLink(props.row.deviceid, props.row.payload)"
                   icon="link"
                   color="grey"
                 >
@@ -182,7 +182,7 @@
                   flat
                   dense
                   size="xs"
-                  @click="deleteATMLink(props.row.id)"
+                  @click="deleteAtmLink(props.row.id)"
                   icon="cancel"
                   color="pink"
                 >
